--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/InjectRuntimeFilter.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/InjectRuntimeFilter.scala
@@ -108,10 +108,8 @@ object InjectRuntimeFilter extends Rule[LogicalPlan] with PredicateHelper with J
       // i.e., the semi-join will be a shuffled join, which is not worthwhile.
       return filterApplicationSidePlan
     }
-    val project = Project(
-      aggregate.output.map(mayWrapWithHash).map(h => Alias(h, h.toString)()), aggregate)
     val filter = InSubquery(Seq(mayWrapWithHash(filterApplicationSideExp)),
-      ListQuery(project, childOutputs = project.output))
+      ListQuery(aggregate, childOutputs = aggregate.output))
     Filter(filter, filterApplicationSidePlan)
   }