### What changes were proposed in this pull request?
Currently, Spark runtime filter supports bloom filter and in subquery filter.
The in subquery filter always execute `Murmur3Hash` before aggregate the join key.

Because the data size before aggregate lager than after, we can only execute `Murmur3Hash` for aggregation and it will reduce the number of calls to `Murmur3Hash` and improve performance.

### Why are the changes needed?
Improve performance for semi-join runtime filter.


### Does this PR introduce _any_ user-facing change?
'No'.
Just update the inner implementation.


### How was this patch tested?
Manually test TPC-DS.
```
spark.sql.optimizer.runtime.bloomFilter.enabled=false
spark.sql.optimizer.runtimeFilter.semiJoinReduction.enabled=true
```

TPC-DS data size: 2TB.
This improvement is valid for below TPC-DS test cases and no regression for other test cases.

| TPC-DS Query   | Before(Seconds)  | After(Seconds)  | Speedup(Percent)  |
|  ----  | ----  | ----  | ----  |
| q23a | 542.03 | 539.06 | 0.55% |
| q23b | 776.7 | 769.74 | 0.90% |
| q24a | 442.25 | 436.75 | 1.26% |
| q24b | 436.16 | 432.86 | 0.76% |
| q50 | 200.92 | 193.36 | 3.91% |
| q64 | 426.73 | 421.03 | 1.35% |
| q67 | 987.79 | 956.82 | 3.24% |
| q93 | 397.26 | 393.9 | 0.85% |
| All TPC-DS | 8392.49 | 8302.56 | 1.08% |