@beliefer 2022-12-21
> is the benchmark result update to update with this new change?

After my test, the results are similar to those of the previous method.